Updated string input related changes

+++ TournamentTreeNode.h
@@ -4,7 +4,7 @@ struct TournamentTreeNode
 {
     RecordStructure element;
     int runId;
-    long long int offsetValueCode = -1;
+    int offsetValueCode = -1;
     // OVC = this.Offset*Domain + (Domain - this.ValueAtOffset)
     int offset = 0;
     // offset = -1 ==> No Offsets computed yet
